added debug to pricing script

+++ AWS/Get-AWSPricingInfo.ps1
@@ -1,4 +1,17 @@
+# Get-AWSPricingInfo.ps1
+# Relies on the AWS CLI, which should be installed and configured
 
+[CmdletBinding()]
+Param(
+    [Parameter(Mandatory=$False)]
+    [String] $Profile,
+    
+    [Parameter(Mandatory=$False)]
+    [String] $Region,
+    
+    [Parameter(Mandatory=$False)]
+    [String] $FiltersFilePath
+)
 
 # identifiers can't start with a number so I added s_ to the front
 enum Ec2Size {
@@ -19,12 +32,45 @@ enum Ec2Size {
 
 ### PART 1 - Use local info to figure out which instance types are valid choices to resize to
 # get relevant instance metadata
-$token = Invoke-RestMethod -Headers @{"X-aws-ec2-metadata-token-ttl-seconds" = "21600"} -Method PUT -Uri http://169.254.169.254/latest/api/token
-$instanceId = Invoke-RestMethod -Headers @{"X-aws-ec2-metadata-token" = $token} -Method GET -Uri http://169.254.169.254/latest/meta-data/instance-id
-$az = Invoke-RestMethod -Headers @{"X-aws-ec2-metadata-token" = $token} -Method GET -Uri http://169.254.169.254/latest/meta-data/placement/availability-zone
-$region = $az.Substring(0, $az.Length - 1)
+try {
+    $token = Invoke-RestMethod -Headers @{"X-aws-ec2-metadata-token-ttl-seconds" = "21600"} -Method PUT -Uri http://169.254.169.254/latest/api/token
+    $instanceId = Invoke-RestMethod -Headers @{"X-aws-ec2-metadata-token" = $token} -Method GET -Uri http://169.254.169.254/latest/meta-data/instance-id
+    $az = Invoke-RestMethod -Headers @{"X-aws-ec2-metadata-token" = $token} -Method GET -Uri http://169.254.169.254/latest/meta-data/placement/availability-zone
+    
+    if($Region) {
+        Write-Host "localhost is running in AWS, -Region flag ignored."
+    }
+    
+    $region = $az.Substring(0, $az.Length - 1)
+}
+catch {
+    Write-Host "Metadata lookup failed. This is normal unless the localhost is in AWS."
+    if(-Not $Region) {
+        Write-Host "USAGE: the -Region flag is required when not running in AWS."
+        exit $False
+    }
+}
 
-$cmd = 'aws ec2 describe-instances --region $region --instance-ids $instanceId'
+if($Profile) {
+    $cmd = 'aws ec2 describe-instances  --profile $Profile --region $region --instance-ids $instanceId'
+}
+else {
+    $cmd = 'aws ec2 describe-instances --region $region --instance-ids $instanceId'
+}
+
+$filters = ""
+if($FiltersFilePath) {
+    $filters = $FiltersFilePath
+}
+else {
+    $filters = "filters.json"
+}
+if(-Not (Test-Path $filters)){
+    Write-Host "Filters file cannot be found. By default we look in the local directory for 'filters.json' but a different location/filename can be specified using the -FiltersFilePath flag."
+    exit $False
+}
+
+Write-Verbose "Gathering AWS metadata for localhost..."
 $instDesc = Invoke-Expression $cmd | ConvertFrom-Json
 $vpcId = $instDesc.Reservations[0].Instances[0].VpcId
 $ebsOptimized = $instDesc.Reservations[0].Instances[0].EbsOptimized
@@ -34,19 +80,20 @@ $typeSize = $instDesc.Reservations[0].Instances[0].InstanceType.Split(".")[1]
 $imageId = $instDesc.Reservations[0].Instances[0].ImageId
 $subnetId = $instDesc.Reservations[0].Instances[0].SubnetId
 $arch = $instDesc.Reservations[0].Instances[0].Architecture
+$customImage = $instDesc.Reservations[0].Instances[0].ProductCodes.Count -gt 0
 
 # get instance type offerings for this zone based on current instance 
+Write-Verbose "Querying AWS for instance types available in this region"
 $types = [System.Collections.ArrayList]@()
-if($vpcId -like "vpc-*") {
-    # current gen vm
-    #$cmd = 'aws ec2 describe-instance-type-offerings --region $region --location-type availability-zone --filters Name=location,Values=$az'
-    $cmd = 'aws ec2 describe-instance-types --region $region --filter Name=current-generation,Values=true'
-    $types = Invoke-Expression $cmd | ConvertFrom-Json
+if($Profile) {
+    # current generation nodes will have a value stored in $vpcId so Values=true in the line below
+    $cmd = 'aws ec2 describe-instance-types --profile $Profile --region $region --filter Name=current-generation,Values=' + ($vpcId -like "vpc-*").ToString().ToLower()
 } else {
-    $cmd = 'aws ec2 describe-instance-types --region $region --filter Name=current-generation,Values=false'
-    $types = Invoke-Expression $cmd | ConvertFrom-Json
+    $cmd = 'aws ec2 describe-instance-types --region $region --filter Name=current-generation,Values=' + ($vpcId -like "vpc-*").ToString().ToLower()
 }
+$types = Invoke-Expression $cmd | ConvertFrom-Json
 
+Write-Verbose "Removing instance types that should not be allowed based on localhost's instance attributes"
 # remove sizes that do not support the current architecture
 $tempTypes = [System.Collections.ArrayList]@()
 $types.InstanceTypes | % { if($_.ProcessorInfo.SupportedArchitectures.Contains($arch)) { $tempTypes.Add($_) > $Null } }
@@ -79,46 +126,64 @@ $types | % {
 $types = $tempTypes
 
 # verify all of these types can be launched in this region
-$tempTypes = [System.Collections.ArrayList]@()
-$types | % {
-    #Write-Host $_.InstanceTYpe
-    #$cmd = 'aws ec2 run-instances --instance-type $_ --dry-run --image-id $imageId --subnet-id $subnetId --region $region'
-    $pinfo = New-Object System.Diagnostics.ProcessStartInfo 
-    $pinfo.FileName = "aws.exe" 
-    $pinfo.Arguments = ("ec2 run-instances --instance-type " + $_.InstanceType + " --dry-run --image-id $imageId --subnet-id $subnetId --region $region")
-    $pinfo.UseShellExecute = $false 
-    $pinfo.CreateNoWindow = $true 
-    $pinfo.RedirectStandardOutput = $true 
-    $pinfo.RedirectStandardError = $true
-    
-    # Create a process object using the startup info
-    $process= New-Object System.Diagnostics.Process 
-    $process.StartInfo = $pinfo
-    
-    # Start the process 
-    $process.Start() | Out-Null 
-    # Wait a while for the process to do something 
-    sleep -Seconds 5 
-    # If the process is still active kill it 
-    if (!$process.HasExited) 
-    { 
-        $process.Kill() 
+# only needed if local instance is using a non base image-id
+if($customImage) {
+    $timeToFinish = $types.Count * 5
+    if($types.Count -lt 50) {
+        $timeToFinish = $types.Count * 10
     }
+    Write-Verbose "localhost was deployed using a Marketplace AMI. Verifying localhost can be launched using allowed sizes. This will take approximately $timeToFinish seconds..."
     
-    $stderr=$process.StandardError.ReadToEnd()
-    if ($stderr.Contains("Request would have succeeded")) 
-    { 
-        #Write-Host "Request would have succeeded"
-        $tempTypes.Add($_) > $Null
+    $tempTypes = [System.Collections.ArrayList]@()
+    $types | % {
+        $pinfo = New-Object System.Diagnostics.ProcessStartInfo 
+        $pinfo.FileName = "aws.exe" 
+        if($Profile) {
+            $pinfo.Arguments = ("ec2 run-instances --profile $Profile --instance-type " + $_.InstanceType + " --dry-run --image-id $imageId --subnet-id $subnetId --region $region")
+        }
+        else {
+            $pinfo.Arguments = ("ec2 run-instances --instance-type " + $_.InstanceType + " --dry-run --image-id $imageId --subnet-id $subnetId --region $region")
+        }
+        $pinfo.UseShellExecute = $false 
+        $pinfo.CreateNoWindow = $true 
+        $pinfo.RedirectStandardOutput = $true 
+        $pinfo.RedirectStandardError = $true
+        
+        # Create a process object using the startup info
+        $process= New-Object System.Diagnostics.Process 
+        $process.StartInfo = $pinfo
+        
+        # Start the process 
+        $process.Start() | Out-Null 
+        # Wait a while for the process to do something 
+        sleep -Seconds 5 
+        # If the process is still active kill it 
+        if (!$process.HasExited) 
+        { 
+            $process.Kill() 
+        }
+        
+        $stderr=$process.StandardError.ReadToEnd()
+        if ($stderr.Contains("Request would have succeeded")) 
+        { 
+            #Write-Host "Request would have succeeded"
+            $tempTypes.Add($_) > $Null
+        }
     }
+    # contains instance types we should allow
+    $types = $tempTypes
 }
-# contains instance types we should allow
-$types = $tempTypes
-
 
 ### PART 2 - Lookup cost info for *all* baseline instance types in the current region
 # find the prices for all valid instance types in this region
-$prices = aws pricing get-products --filters file://filters.json --format-version aws_v1 --service-code AmazonEC2 --region us-east-1 | ConvertFrom-Json
+Write-Verbose "Looking up pricing info for base instance sizes"
+if($Profile) {
+    $cmd = 'aws pricing get-products --profile $Profile --filters file://$filters --format-version aws_v1 --service-code AmazonEC2 --region us-east-1'
+}
+else {
+    $cmd = 'aws pricing get-products --filters file://$filters --format-version aws_v1 --service-code AmazonEC2 --region us-east-1'
+}
+$prices = Invoke-Expression $cmd | ConvertFrom-Json
 $priceList = $prices.PriceList | ConvertFrom-Json
 
 # store all valid cost values in a hashtable of <instanceType, cost> pairs
@@ -143,15 +208,16 @@ $priceList | % {
 
 
 ### Part 3 - output only allowed instance types and their annual costs
+Write-Host "Allowed instance types and their annual costs:"
 $output = [hashtable]@{}
  $types | % {
-    #$output.Add($_.InstanceType, [math]::Round($ht[$_.InstanceType] * 8766, 2)) > $Null
+    $output.Add($_.InstanceType, [math]::Round($ht[$_.InstanceType] * 8766, 2)) > $Null
 }
 $output.GetEnumerator() | Sort -Property Name
+Write-Host "Annual cost for local VM is currently " + $output[($typeFamily + "." + $typeFamily)] + " USD.`n"
 
-<### debug - the below code will display the full pricelist for this region, sorted.
-$ht.Keys | % {
-    $output.Add($_, [math]::Round($ht[$_] * 8766, 2)) > $Null
+### debug - the below code will display the full pricelist for this region, sorted.
+Write-Verbose "Hourly costs for all possible instance types in this region"
+if($PSCmdlet.MyInvocation.BoundParameters["Verbose"].IsPresent) {
+    $ht.GetEnumerator() | Sort -Property Name
 }
-$output.GetEnumerator() | Sort -Property Name
-#>
